add lazy clone and virtualizing objects
Begin adding support for the ability to have a lazy clone that does not
download all objects at clone time. This is to support "virtualizing"
very large repos where the number and/or size of the objects makes it
impractical to download them all at clone time.  Currently, a new
"core.virtualizeobjects" flag has to be set to enable this capability.

This patch series adds support for a having a background process that
can retrieve missing blobs "on demand" as they are needed and then store
them in the regular object store as loose objects or pack files. It is
spawned the first time a missing blob is encounterd and continues
running in the background waiting for additional requests until the
current git command completes.  This model eliminiates the overhead of
having to spawn a separate process for each missing object.

 - Refactored the filter process code into a generic reusable implementation
 - Added calls to a read-object process when objects can not be found
 - Wrote a read-object background process test script, test cases, documentation

Signed-off-by: Ben Peart <benpeart@microsoft.com>
Signed-off-by: Christian Couder <chriscool@tuxfamily.org>