odb-helper: add odb_helper_init() to send "init" command
Let's add an odb_helper_init() function to send an "init"
command to the helpers. This "init" command is especially
useful to get the capabilities that are supported by the
helpers.

So while at it, let's also add a parse_capabilities()
function to parse them and a supported_capabilities
variable in struct odb_helper to store them.

Signed-off-by: Christian Couder <chriscool@tuxfamily.org>