odb-remote: add 'get_direct' support
This implements the 'get_direct' capability/instruction that makes
it possible for odb helper scripts to pass blobs to Git
by directly writing them as loose objects files.

It is better to call this a "direct" mode rather than a "fault-in"
mode as we could have the same kind of mechanism to "put" objects
into an odb remote, where the odb helper would access blobs it
wants to send to an odb remote directly from files, but it
would be strange to call that a fault-in mode too.

Signed-off-by: Christian Couder <chriscool@tuxfamily.org>