t0021/rot13-filter: add capability functions
Add functions to help read and write capabilities.
These functions will be reused in following patches.

 One more thing that is more noteworthy (read: do not forget to
describe it in the proposed log message) is that the original used
to require capabilities to come in a fixed order.

The new code allows these capabilities to be declared in any order,
it even allows duplicates (intended? shouldn't we be flagging it as
an error?), the helper can require a set of capabilities we do want
to see and fail if the remote doesn't declare any one of them
(good).

It does not check if the remote declares any capability we do not
know about (intended? the original noticed this situation and error
out---shouldn't the more generalized helper that is meant to be
reusable allow us to do so, too?).

        Side note: my answer to the last question is "it is OK and
        even desirable to ignore the fact that they claim to support
        a capability we do not know about", but I may be mistaken.
        The reasoning and the conclusion that is consistent with
        what the code does should be described in any case.

Signed-off-by: Christian Couder <chriscool@tuxfamily.org>