odb-helper: add 'enum odb_helper_type'
As there will be different kinds of helpers, let's add
an "enum odb_helper_type" to tell between the different
kinds.

Let's add a field with this type in "struct odb_helper",
and set it when reading the config file.

While at it let's also make it possible to find an helper
of a specific kind by adding a 'type' parameter in the
find_odb_helper() function.

Signed-off-by: Christian Couder <chriscool@tuxfamily.org>