pack-objects: move tree_depth into 'struct packing_data'
This reduces the size of 'struct object_entry' and therefore
makes packing objects more efficient.

This also renames cmp_tree_depth() into tree_depth_compare(),
as it is more modern to have the name of the compare functions
end with "compare".

QSORT() is used instead of qsort() to further modernize the
code.

Signed-off-by: Christian Couder <chriscool@tuxfamily.org>