promisor-remote: use repository_format_partial_clone
A remote specified using the extensions.partialClone config
option should be considered a promisor remote too.

This remote should be at the end of the promisor remote list,
so that it is used only if objects have not been found in other
remotes.

If we are using promisor remotes other than the origin, it is
because these other promisor remotes are likely to be better
for some reason (maybe they are closer or faster for some kind
of objects) than the origin, so it's logical that we try to use
them first.

Signed-off-by: Christian Couder <chriscool@tuxfamily.org>