ewah/bitmap: introduce bitmap_word_alloc()
In a following patch we will need to allocate a variable
number of bitmap words, instead of always 32, so let's add
bitmap_word_alloc() for this purpose.

We will also always access at least one word for each bitmap,
so we want to make sure that at least one is always
allocated.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Christian Couder <chriscool@tuxfamily.org>