ewah/bitmap: always allocate at least one word
In a following patch we will allocate a variable number
of words in some bitmaps. When iterating over the words we
will need to make sure we always make forward progress.

We will always access at least one word for each bitmap
so we want to make sure that at least one is always
allocated.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Christian Couder <chriscool@tuxfamily.org>