promisor-remote: refactor to get rid of 'struct strvec'
In a following commit, we will use the new 'promisor-remote' protocol
capability introduced by d460267613 (Add 'promisor-remote' capability
to protocol v2, 2025-02-18) to pass and process more information
about promisor remotes than just their name and url.

For that purpose, we will need to store information about other
fields, especially information that might or might not be available
for different promisor remotes. Unfortunately using 'struct strvec',
as we currently do, to store information about the promisor remotes
with one 'struct strvec' for each field like "name" or "url" does not
scale easily in that case.

Let's refactor this and introduce a new 'struct promisor_info' linked
list that contains a 'struct string_list fields'. This string_list
stores the field names, like "name" and "url", in the 'string' member
of its items, and the field values in the 'util' member of its items.

Except for "name", each "<field_name>/<field_value>" pair should
correspond to a "remote.<name>.<field_name>" config variable set to
<field_value> where "<name>" is a promisor remote name.

Previously in Git, the part after the last dot in a configuration
variable key, for example "c" for "a.b.c", was called the "variable
name part" of a configuration key. It would be very confusing to use
"variable name part" or some similar terms in the context of the
'promisor-remote' protocol though, so let's forget about it, and just
use "field", "field name" and "field value" instead.

Signed-off-by: Christian Couder <chriscool@tuxfamily.org>